[REFACTOR/#113]: 공동구매 오픈 submit함수 커스텀 훅화
